WP-72 WP-83 WP-90 Implemented two pages, improved HTML, allowed user input for URLs
