+++ NodeGraphWeb/Pages/Build.cshtml
@@ -1,13 +1,18 @@
 ﻿@page
-@model NodeGraphWeb.Pages.Index1Model
+@model NodeGraphWeb.Pages.GraphModel2
 @{
     ViewData["Title"] = "Build Graph";
 }
 
 <h1>Graph Visualization</h1>
 
+<form method="post">
+    <input type="text" name="urlInput" placeholder="Enter a URL" required>
+    <button type="submit">Build Graph</button>
+</form>
+
 <div id="container">
-    <div id="graphContainer"></div>
+    <div id="graphContainer2"></div>
 </div>
 <div id="tooltip" style="position: absolute; visibility: hidden; background-color: #fff; border: 1px solid #000; padding: 5px; pointer-events: none;"></div>
 
@@ -19,7 +24,7 @@
         border: 2px solid black;
     }
 
-    #graphContainer {
+    #graphContainer2 {
         width: 100%;
         height: 900px;
         position: relative;
@@ -49,8 +54,10 @@
 </style>
 
 <script src="https://d3js.org/d3.v6.min.js"></script>
-<script src="~/js/graph.js"></script> <!-- Link to your new JavaScript file -->
 
 <script>
-    var graphData = @Html.Raw(Model.GraphJsonData); // Inject the JSON data into the page
+    // Assign the serialized graph data to a JavaScript variable
+    var graphData = @Html.Raw(Model.GraphJsonData);
 </script>
+
+<script src="~/js/buildgraph.js"></script>
