+++ NodeGraphWeb/Pages/Build.cshtml.cs
@@ -1,30 +1,34 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.Msagl.Drawing;
-using NodeGraphWeb.Services;
+using NodeGraphWeb.Services2;
 
 namespace NodeGraphWeb.Pages
 {
     public class GraphModel2 : PageModel
     {
-        private readonly GraphService _graphService;
+        private readonly GraphServiceBuild _graphService;
 
         public string? GraphJsonData { get; private set; }
 
-        public GraphModel2(GraphService graphService2)
+        public GraphModel2(GraphServiceBuild graphService)
         {
-            _graphService2 = graphService2;
+            _graphService = graphService;
         }
 
-        public void OnGet()
-        {
-            //Fetch the graph data in JSON format from the database
-            GraphJsonData = _graphService.GetGraphDataFromDatabase();
+        [BindProperty]
+        public string UrlInput { get; set; }
 
-            /* // previous code for fetching directly from backend
-            var graph = _graphService.BuildGraph("https://github.com");
+        public IActionResult OnPost()
+        {
+            if (!ModelState.IsValid)
+            {
+                return Page();
+            }
+
+            var graph = _graphService.BuildGraph(UrlInput);
             GraphJsonData = _graphService.ConvertGraphToJson(graph);
-            */
+            return Page();
         }
     }
 }
