+++ NodeGraphWeb/Pages/Shared/_Layout.cshtml
@@ -28,6 +28,9 @@
                         <li class="nav-item">
                             <a class="nav-link text-dark" asp-area="" asp-page="/Graph">Graph</a>
                         </li>
+                        <li class="nav-item">
+                            <a class="nav-link text-dark" asp-area="" asp-page="/Build">Build Graph</a>
+                        </li>
                     </ul>
                 </div>
             </div>
